+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Unreleased
 
+- *Breaking:* The feature `defmt-03` has been renamed `defmt` and has been updated to 1.0.
+
 ## 5.0.1 27-08-25
 
 - Fixed the `Value` implementation for `Option<T>` so it now serializes and deserializes properly